@@ -2,14 +2,16 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-using Bet.AspNetCore.EvenGrid.Webhooks;
+using Bet.AspNetCore.EventGrid.Abstractions;
+using Bet.AspNetCore.EventGrid.Abstractions.Models;
 using Bet.AspNetCore.EventGrid.WebApp.Events;
 using Bet.AspNetCore.EventGrid.WebApp.Services;
+
 using Microsoft.Extensions.Logging;
 
 namespace Bet.AspNetCore.EventGrid.WebApp.Handler
 {
-    public class EmployeeWebhook : IEventGridWebhook<EmployeeCreatedEvent>
+    public class EmployeeWebhook : IWebhook<EmployeeCreatedEvent>
     {
         private readonly ILogger<EmployeeWebhook> _logger;
         private readonly IOperationService _service;
@@ -20,7 +22,7 @@ namespace Bet.AspNetCore.EventGrid.WebApp.Handler
             _service = service ?? throw new System.ArgumentNullException(nameof(service));
         }
 
-        public Task<EventGridWebHookResult> ProcessEventAsync(EmployeeCreatedEvent @event, CancellationToken cancellationToken = default)
+        public Task<WebHookResult> ProcessEventAsync(EmployeeCreatedEvent @event, CancellationToken cancellationToken = default)
         {
             _logger.LogInformation($"Transient Id: {_service.TransientOperation.OperationId}{Environment.NewLine}" +
                 $"Scoped Id: {_service.ScopedOperation.OperationId}{Environment.NewLine}" +
@@ -28,10 +30,9 @@ namespace Bet.AspNetCore.EventGrid.WebApp.Handler
                 $"Singleton Instance Id: {_service.SingletonInstanceOperation.OperationId}{Environment.NewLine}");
 
             // var data = @event as EmployeeCreatedEvent;
-
             _logger.LogInformation("Executing: {data} ", @event);
 
-            return Task.FromResult<EventGridWebHookResult>(null);
+            return Task.FromResult<WebHookResult>(null);
         }
     }
 }
